Fixed some more warnings.

+++ extras/JuceDemo/JuceLibraryCode/BinaryData.cpp
@@ -2147,7 +2147,7 @@ const char* getNamedResource (const char* resourceNameUTF8, int& numBytes) throw
     unsigned int hash = 0;
     if (resourceNameUTF8 != 0)
         while (*resourceNameUTF8 != 0)
-            hash = 31 * hash + *resourceNameUTF8++;
+            hash = 31 * hash + (unsigned int) *resourceNameUTF8++;
 
     switch (hash)
     {
