+++ modules/juce_gui_basics/windows/juce_TooltipWindow.cpp
@@ -183,7 +183,7 @@ void TooltipWindow::timerCallback()
         // appear after a timeout..
         if (newTip.isNotEmpty()
              && newTip != tipShowing
-             && now > lastCompChangeTime + millisecondsBeforeTipAppears)
+             && now > lastCompChangeTime + (unsigned int) millisecondsBeforeTipAppears)
         {
             showFor (newTip);
         }
